Finished Pygame Tutorial
I finished the pygame tutorial on YouTube. I was kind of annoyed that the instructor wasn't using OOP because that would have simplified a few things and made the game better, so I did my version of the exact same game and improved it with a few classes. It's really basic but I think it's pretty good. I now want to make a game with pygame and see what I can do on my own.