Started a FreeCodeCamp tutorial on Tkinter
I started a free code camp tutorial about using python's Tkinter, but I'm going to put it on hold since I don't feel that the instructor is going as fast as I'd like.

I might come back to it, but I'm going to work on a different project in the mean time.