Finished Endgame and DNF Assignments
This update includes the code for the Endgame assignment for CIS 203. Endgame added a time countdown, an alert dialog, and a high score check.

The CIS 206 DNF assignment was to read in a table of values and calculate the Disjunctive Normal Form of the binary equation in the table. I also did the extra credit to calculate the CNF of the table.